# backend/src/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Docker Web Desktop API",
    "description": "API documentation for Docker Web Desktop (auto-generated)",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/api",
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/images/": {
      "get": {
        "tags": [
          "Images"
        ],
        "description": "",
        "parameters": [
          {
            "name": "source",
            "in": "query",
            "type": "string"
          },
          {
            "name": "wslDistro",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/images/{id}": {
      "delete": {
        "tags": [
          "Images"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "force",
            "in": "query",
            "type": "string"
          },
          {
            "name": "source",
            "in": "query",
            "type": "string"
          },
          {
            "name": "wslDistro",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/volumes/": {
      "get": {
        "tags": [
          "Volumes"
        ],
        "description": "",
        "parameters": [
          {
            "name": "source",
            "in": "query",
            "type": "string"
          },
          {
            "name": "wslDistro",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/volumes/{name}": {
      "delete": {
        "tags": [
          "Volumes"
        ],
        "description": "",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "force",
            "in": "query",
            "type": "string"
          },
          {
            "name": "source",
            "in": "query",
            "type": "string"
          },
          {
            "name": "wslDistro",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/networks/": {
      "get": {
        "tags": [
          "Networks"
        ],
        "description": "",
        "parameters": [
          {
            "name": "source",
            "in": "query",
            "type": "string"
          },
          {
            "name": "wslDistro",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/networks/{id}": {
      "delete": {
        "tags": [
          "Networks"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "force",
            "in": "query",
            "type": "string"
          },
          {
            "name": "source",
            "in": "query",
            "type": "string"
          },
          {
            "name": "wslDistro",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/services/": {
      "get": {
        "tags": [
          "Services"
        ],
        "description": "",
        "parameters": [
          {
            "name": "source",
            "in": "query",
            "type": "string"
          },
          {
            "name": "wslDistro",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/services/{id}": {
      "get": {
        "tags": [
          "Services"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "source",
            "in": "query",
            "type": "string"
          },
          {
            "name": "wslDistro",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "tags": [
          "Services"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "source",
            "in": "query",
            "type": "string"
          },
          {
            "name": "wslDistro",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/wsl/": {
      "get": {
        "tags": [
          "WSL"
        ],
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/wsl/{id}": {
      "get": {
        "tags": [
          "WSL"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  }
}